іг

+++ src/services/contacts.js
@@ -1,68 +1,48 @@
-import { ContactsCollection } from "../db/models/contact.js";
-import { calculatePaginationData } from "../utils/calculatePaginationData.js";
-import { SORT_ORDER } from "../constants/index.js";
+import { ContactsCollection } from '../db/models/contact.js';
+import { calculatePaginationData } from '../utils/calculatePaginationData.js';
 
+export const getAllContacts = async ({ page, perPage, sortOrder, sortBy, userId }) => {
+    const limit = perPage;
+    const skip = page > 0 ? (page - 1) * perPage : 0;
 
-export const getAllContacts = async ({
-  page = 1,
-  perPage = 10,
-  sortOrder = SORT_ORDER.ASC,
-  sortBy = '_id',
-  userId,
-  
-}) => {
-  const limit = perPage;
-  const skip = (page - 1) * perPage;
+    const contactsQuery = ContactsCollection.find({ userId });
 
-  const contactsQuery = ContactsCollection.find({ userId });
-    const contactsCount = await ContactsCollection.find({ userId }).merge(contactsQuery).countDocuments();
-  const contacts = await contactsQuery.skip(skip).limit(limit).sort({ [sortBy]: sortOrder }).exec();
-  const paginationData = calculatePaginationData(contactsCount, perPage, page);
+    const [total, contacts] = await Promise.all([
+        ContactsCollection.countDocuments(contactsQuery),
+        contactsQuery
+            .sort({ [sortBy]: sortOrder })
+            .skip(skip)
+            .limit(limit),
+    ]);
+
+    const paginationData = calculatePaginationData(total, perPage, page);
+
     return {
-    data: contacts,
-    ...paginationData,
-  };
-
+        data: contacts,
+        ...paginationData,
+    };
 };
 
 export const getContactById = async (contactId, userId) => {
-  const contact = await ContactsCollection.findOne({
-    _id: contactId,
-    userId: userId,
-  });
+    const contact = await ContactsCollection.findOne({ _id: contactId, userId });
     return contact;
 };
 
-
-export const createContact = async (payload) => {
-  const contact = await ContactsCollection.create(payload);
-  return contact;
+export const createContact = async (newContact) => {
+    const contact = await ContactsCollection.create(newContact);
+    return contact;
 };
 
 export const deleteContact = async (contactId, userId) => {
-   const contact = await ContactsCollection.findOneAndDelete({
-     _id: contactId,
-     userId,
-  });
-
-  return contact;
+    const contact = await ContactsCollection.findOneAndDelete({ _id: contactId, userId });
+    return contact;
 };
 
-export const updateContact = async (contactId, userId, payload, options = {}) => {
-const rawResult = await ContactsCollection.findOneAndUpdate(
-    { _id: contactId, userId },
-    payload,
-    {
-      new: true,
-      includeResultMetadata: true,
-      ...options,
-    },
-  );
-
-  if (!rawResult || !rawResult.value) return null;
-
-  return {
-    contact: rawResult.value,
-    isNew: Boolean(rawResult?.lastErrorObject?.upserted),
-  };
+export const updateContact = async (contactId, updateContact, userId, options = {}) => {
+    const contact = await ContactsCollection.findOneAndUpdate(
+        { _id: contactId, userId },
+        updateContact,
+        { new: true, ...options },
+    );
+    return contact;
 };
